[Synthetics] add default email recovery message (#154862)
## Summary

Resolves https://github.com/elastic/kibana/issues/153891
<img width="1334" alt="Screen Shot 2023-04-18 at 7 50 20 PM"
src="https://user-images.githubusercontent.com/11356435/232928889-09de824f-14ef-426d-8732-6da7fc12eb37.png">

Down email

Recovered email
<img width="1339" alt="Screen Shot 2023-04-18 at 7 50 30 PM"
src="https://user-images.githubusercontent.com/11356435/232928874-4e7c0cbb-b7ec-4684-b0d2-88ec993697df.png">

Please note the `recoverd` typo in the screenshot has been addressed in
commits
[7c4040f](https://github.com/elastic/kibana/pull/154862/commits/7c4040fda316f0565eb1ea429417b0fa91f6424d)
and
[56f637b](https://github.com/elastic/kibana/pull/154862/commits/56f637b82edf81b7ab6dafd54c197f1b945c487b)

### Testing
This PR is extremely tricky to test. I have deployed this branch to my
own cloud account. This ensures that I can control the allowlist setting
for emails on cloud. If you want, I can give you the cloud cluster where
this branch is hosted and if you give me your email I can add you to my
allowlist settings for you to test.

The cloud cluster is current on commit
[f889500](https://github.com/elastic/kibana/pull/154862/commits/f889500dc2b217fefb1a02a9e3c15b7cad8217cb)

(cherry picked from commit 0b71c7458df0d78869668dbe38c7b11757da3193)
