+++ x-pack/plugins/synthetics/common/rules/alert_actions.test.ts
@@ -9,7 +9,7 @@ import { populateAlertActions } from './alert_actions';
 import { ActionConnector } from './types';
 import { MONITOR_STATUS } from '../constants/uptime_alerts';
 import { MONITOR_STATUS as SYNTHETICS_MONITOR_STATUS } from '../constants/synthetics_alerts';
-import { MonitorStatusTranslations } from '../translations';
+import { MonitorStatusTranslations } from './legacy_uptime/translations';
 import { SyntheticsMonitorStatusTranslations } from './synthetics/translations';
 
 describe('Legacy Alert Actions factory', () => {
@@ -33,6 +33,7 @@ describe('Legacy Alert Actions factory', () => {
         defaultActionMessage: MonitorStatusTranslations.defaultActionMessage,
         defaultRecoveryMessage: MonitorStatusTranslations.defaultRecoveryMessage,
         defaultSubjectMessage: MonitorStatusTranslations.defaultSubjectMessage,
+        defaultRecoverySubjectMessage: MonitorStatusTranslations.defaultRecoverySubjectMessage,
       },
       isLegacy: true,
     });
@@ -60,6 +61,70 @@ describe('Legacy Alert Actions factory', () => {
     ]);
   });
 
+  it('generate expected action for email', async () => {
+    const resp = populateAlertActions({
+      groupId: MONITOR_STATUS.id,
+      defaultActions: [
+        {
+          actionTypeId: '.email',
+          group: 'xpack.uptime.alerts.actionGroups.monitorStatus',
+          params: {
+            dedupKey: 'always-downxpack.uptime.alerts.actionGroups.monitorStatus',
+            eventAction: 'trigger',
+            severity: 'error',
+            summary: MonitorStatusTranslations.defaultActionMessage,
+          },
+          id: 'f2a3b195-ed76-499a-805d-82d24d4eeba9',
+        },
+      ] as unknown as ActionConnector[],
+      translations: {
+        defaultActionMessage: MonitorStatusTranslations.defaultActionMessage,
+        defaultRecoveryMessage: MonitorStatusTranslations.defaultRecoveryMessage,
+        defaultSubjectMessage: MonitorStatusTranslations.defaultSubjectMessage,
+        defaultRecoverySubjectMessage: MonitorStatusTranslations.defaultRecoverySubjectMessage,
+      },
+      isLegacy: true,
+      defaultEmail: {
+        to: ['test@email.com'],
+      },
+    });
+    expect(resp).toEqual([
+      {
+        group: 'recovered',
+        id: 'f2a3b195-ed76-499a-805d-82d24d4eeba9',
+        params: {
+          bcc: [],
+          cc: [],
+          kibanaFooterLink: {
+            path: '',
+            text: '',
+          },
+          message:
+            'Alert for monitor {{context.monitorName}} with url {{{context.monitorUrl}}} from {{context.observerLocation}} has recovered',
+          subject:
+            'Monitor {{context.monitorName}} with url {{{context.monitorUrl}}} has recovered',
+          to: ['test@email.com'],
+        },
+      },
+      {
+        group: 'xpack.uptime.alerts.actionGroups.monitorStatus',
+        id: 'f2a3b195-ed76-499a-805d-82d24d4eeba9',
+        params: {
+          bcc: [],
+          cc: [],
+          kibanaFooterLink: {
+            path: '',
+            text: '',
+          },
+          message:
+            'Monitor {{context.monitorName}} with url {{{context.monitorUrl}}} from {{context.observerLocation}} {{{context.statusMessage}}} The latest error message is {{{context.latestErrorMessage}}}, checked at {{context.checkedAt}}',
+          subject: 'Monitor {{context.monitorName}} with url {{{context.monitorUrl}}} is down',
+          to: ['test@email.com'],
+        },
+      },
+    ]);
+  });
+
   it('generate expected action for index', async () => {
     const resp = populateAlertActions({
       groupId: MONITOR_STATUS.id,
@@ -80,6 +145,7 @@ describe('Legacy Alert Actions factory', () => {
         defaultActionMessage: MonitorStatusTranslations.defaultActionMessage,
         defaultRecoveryMessage: MonitorStatusTranslations.defaultRecoveryMessage,
         defaultSubjectMessage: MonitorStatusTranslations.defaultSubjectMessage,
+        defaultRecoverySubjectMessage: MonitorStatusTranslations.defaultRecoverySubjectMessage,
       },
       isLegacy: true,
     });
@@ -141,6 +207,7 @@ describe('Legacy Alert Actions factory', () => {
         defaultActionMessage: MonitorStatusTranslations.defaultActionMessage,
         defaultRecoveryMessage: MonitorStatusTranslations.defaultRecoveryMessage,
         defaultSubjectMessage: MonitorStatusTranslations.defaultSubjectMessage,
+        defaultRecoverySubjectMessage: MonitorStatusTranslations.defaultRecoverySubjectMessage,
       },
     });
     expect(resp).toEqual([
@@ -189,6 +256,8 @@ describe('Alert Actions factory', () => {
         defaultActionMessage: SyntheticsMonitorStatusTranslations.defaultActionMessage,
         defaultRecoveryMessage: SyntheticsMonitorStatusTranslations.defaultRecoveryMessage,
         defaultSubjectMessage: SyntheticsMonitorStatusTranslations.defaultSubjectMessage,
+        defaultRecoverySubjectMessage:
+          SyntheticsMonitorStatusTranslations.defaultRecoverySubjectMessage,
       },
     });
     expect(resp).toEqual([
@@ -235,6 +304,8 @@ describe('Alert Actions factory', () => {
         defaultActionMessage: SyntheticsMonitorStatusTranslations.defaultActionMessage,
         defaultRecoveryMessage: SyntheticsMonitorStatusTranslations.defaultRecoveryMessage,
         defaultSubjectMessage: SyntheticsMonitorStatusTranslations.defaultSubjectMessage,
+        defaultRecoverySubjectMessage:
+          SyntheticsMonitorStatusTranslations.defaultRecoverySubjectMessage,
       },
     });
     expect(resp).toEqual([
@@ -295,6 +366,8 @@ describe('Alert Actions factory', () => {
         defaultActionMessage: SyntheticsMonitorStatusTranslations.defaultActionMessage,
         defaultRecoveryMessage: SyntheticsMonitorStatusTranslations.defaultRecoveryMessage,
         defaultSubjectMessage: SyntheticsMonitorStatusTranslations.defaultSubjectMessage,
+        defaultRecoverySubjectMessage:
+          SyntheticsMonitorStatusTranslations.defaultRecoverySubjectMessage,
       },
     });
     expect(resp).toEqual([
@@ -316,6 +389,71 @@ describe('Alert Actions factory', () => {
           eventAction: 'trigger',
           severity: 'error',
           summary: SyntheticsMonitorStatusTranslations.defaultActionMessage,
+        },
+      },
+    ]);
+  });
+
+  it('generate expected action for email action connector', async () => {
+    const resp = populateAlertActions({
+      groupId: SYNTHETICS_MONITOR_STATUS.id,
+      defaultActions: [
+        {
+          actionTypeId: '.email',
+          group: 'xpack.synthetics.alerts.actionGroups.monitorStatus',
+          params: {
+            dedupKey: 'always-downxpack.uptime.alerts.actionGroups.monitorStatus',
+            eventAction: 'trigger',
+            severity: 'error',
+            summary:
+              'Monitor {{context.monitorName}} with url {{{context.monitorUrl}}} from {{context.observerLocation}} {{{context.statusMessage}}} The latest error message is {{{context.latestErrorMessage}}}',
+          },
+          id: 'f2a3b195-ed76-499a-805d-82d24d4eeba9',
+        },
+      ] as unknown as ActionConnector[],
+      defaultEmail: {
+        to: ['test@email.com'],
+      },
+      translations: {
+        defaultActionMessage: SyntheticsMonitorStatusTranslations.defaultActionMessage,
+        defaultRecoveryMessage: SyntheticsMonitorStatusTranslations.defaultRecoveryMessage,
+        defaultSubjectMessage: SyntheticsMonitorStatusTranslations.defaultSubjectMessage,
+        defaultRecoverySubjectMessage:
+          SyntheticsMonitorStatusTranslations.defaultRecoverySubjectMessage,
+      },
+    });
+    expect(resp).toEqual([
+      {
+        group: 'recovered',
+        id: 'f2a3b195-ed76-499a-805d-82d24d4eeba9',
+        params: {
+          bcc: [],
+          cc: [],
+          kibanaFooterLink: {
+            path: '',
+            text: '',
+          },
+          message:
+            'The alert for the monitor {{context.monitorName}} checking {{{context.monitorUrl}}} from {{context.locationName}} is no longer active: {{context.recoveryReason}}.',
+          subject:
+            'The monitor {{context.monitorName}} checking {{{context.monitorUrl}}} has recovered.',
+          to: ['test@email.com'],
+        },
+      },
+      {
+        group: 'xpack.synthetics.alerts.actionGroups.monitorStatus',
+        id: 'f2a3b195-ed76-499a-805d-82d24d4eeba9',
+        params: {
+          bcc: [],
+          cc: [],
+          kibanaFooterLink: {
+            path: '',
+            text: '',
+          },
+          message:
+            'The monitor {{context.monitorName}} checking {{{context.monitorUrl}}} from {{context.locationName}} last ran at {{context.checkedAt}} and is {{{context.status}}}. The last error received is: {{{context.lastErrorMessage}}}.',
+          subject: 'The monitor {{context.monitorName}} checking {{{context.monitorUrl}}} is down.',
+          to: ['test@email.com'],
         },
       },
     ]);
