+++ x-pack/plugins/synthetics/public/legacy_uptime/lib/alert_types/duration_anomaly.tsx
@@ -14,7 +14,7 @@ import { ObservabilityRuleTypeModel } from '@kbn/observability-plugin/public';
 import { AlertTypeInitializer } from '.';
 import { getMonitorRouteFromMonitorId } from '../../../../common/utils/get_monitor_url';
 import { CLIENT_ALERT_TYPES } from '../../../../common/constants/uptime_alerts';
-import { DurationAnomalyTranslations } from '../../../../common/translations';
+import { DurationAnomalyTranslations } from '../../../../common/rules/legacy_uptime/translations';
 
 const { defaultActionMessage, defaultRecoveryMessage, description } = DurationAnomalyTranslations;
 const DurationAnomalyAlert = React.lazy(() => import('./lazy_wrapper/duration_anomaly'));
