+++ x-pack/plugins/synthetics/server/legacy_uptime/lib/alerts/tls.ts
@@ -21,7 +21,7 @@ import { CLIENT_ALERT_TYPES, TLS } from '../../../../common/constants/uptime_ale
 import { DYNAMIC_SETTINGS_DEFAULTS } from '../../../../common/constants';
 import { Cert, CertResult } from '../../../../common/runtime_types';
 import { commonStateTranslations, tlsTranslations } from './translations';
-import { TlsTranslations } from '../../../../common/translations';
+import { TlsTranslations } from '../../../../common/rules/legacy_uptime/translations';
 
 import { savedObjectsAdapter } from '../saved_objects/saved_objects';
 import { UptimeEsClient } from '../lib';
