+++ x-pack/plugins/synthetics/server/routes/default_alerts/status_alert_service.ts
@@ -114,6 +114,8 @@ export class StatusAlertService {
         defaultActionMessage: SyntheticsMonitorStatusTranslations.defaultActionMessage,
         defaultRecoveryMessage: SyntheticsMonitorStatusTranslations.defaultRecoveryMessage,
         defaultSubjectMessage: SyntheticsMonitorStatusTranslations.defaultSubjectMessage,
+        defaultRecoverySubjectMessage:
+          SyntheticsMonitorStatusTranslations.defaultRecoverySubjectMessage,
       },
     });
   }
