+++ x-pack/test/functional_with_es_ssl/apps/discover_ml_uptime/uptime/simple_down_alert.ts
@@ -6,7 +6,7 @@
  */
 
 import expect from '@kbn/expect';
-import { MonitorStatusTranslations } from '@kbn/synthetics-plugin/common/translations';
+import { MonitorStatusTranslations } from '@kbn/synthetics-plugin/common/rules/legacy_uptime/translations';
 import { FtrProviderContext } from '../../../ftr_provider_context';
 import { deleteUptimeSettingsObject } from '../../../../functional/apps/uptime';
